requirements.txt
uv
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-openai
openai
python-dotenv
datasets
ipykernel